Refactor power actions into hook
